feat(server): add database connection parameter retrieval API
Added a new API endpoint to the CodeGeneratorController to retrieve thedatabase connection parameters. This API will facilitate the display of
connected database information on the front-end, improving user
usability and transparency in the code generation process.

The `getConnectDatabase` method fetches the connection parameters from
the session attribute and returns them in the response. This change
is part of the ongoing code generator service enhancements.